feat(gallatin): move task type definitions to library
