+++ libs/interfaces/src/grpc/index.ts
@@ -1,0 +1,1 @@
+export * from 'tasks.ts';
